Working Temperature Generator
It is working and it has a lot of potential. So far it gives the current
temperature according to the time of the day and has a variation between
day and night. 
It has a feature for giving a temperature status, latter that can be
modified to do an auto share if temperature at risk.