Chart Class Java FX
Chart Class uses Java Fx to create a chart. The implications of using
java fx is that the compatibility with swing is very different,
therefore there must be many workarounds to run FX on swing, due to all
of the program using swing. Once I defeat the learning curve, it will be
'easier' to create beautiful graphs. Thus, JavaFX is included on Java
SDK rather than JChartsFree that has to be downloaded and configured on
each machine.