Added Colors to Heartbeat Chart
It took a looooooooot of time to add colors to the chart, but it looks
neat, the next step is to mess up with the .css file (that's new by the
way). Also the custom jpanel that has the chart is working so now I have
to link it to the action listener, and voila!