Custom View Screen Navigation
Each screen now references to the correct graph type(yet to be
implemented). As right now only the heartbeat is working but it finally
displays both weekly and monthly charts. 

Added an interface called Statistics for the individual generators,
therefore it can be reused four times. JFreeChart is one step to be
deleted because as right now JavaFX charts are working great.