Created Abstract Classes for Generators
I removed the interface because it was not working, and after visiting
my professor I understood that the Interfaces are designed at the
beginning of the project and right now there was not a use for them.
Instead I created two abstract classes that do the job well. They are
not essential to the program, but it demonstrates knowledge of how
abstract classes work and they make the code somewhat cleaner. One thing
that I liked about abstract methods is that they are required in the
inherited classes, therefore you cannot forget to add them.

*Added charts for temperature which created some bugs
*Added Sleep Generator. It has no real information whatsoever.

*BUG1: Temperature month chart not showing, well it is showing but links
to the month chart, which sucks.
*BUG2: COLORS IN CHARTS ARE NOT SHOWING!!!!!! GGGGRRRRR!! I spent a lot
of time with that and suddenly they are not working. CSS do work but it
will be a pain to color everything on CSS.