Sleep Generator and Chart
Both are working.
Generator is pretty lame because it does not have many maths, and the
random numbers are meh.
Chart is working but throws many errors because it expects more values
than just one array. There are two possible workarounds, or either pass
not extend the Abstract Charts and do it everything on the sleepChart,
or break your head and make it work with the current extension. CSS
seems like a good idea for this chart.