Database Resources Test Files
As right now the database is able to identify repeating strings and integers between a new User and a User that is already part of the database. The next step is to be able to traverse the database, change the user, logi-in, log-out. Finally to save the information in a txt file, or a real database would be awesome.