Database FIXED
Finally the database is working. The database class follows the
singleton design pattern which works perfectly because before that there
was a __global__ database called from everywhere. Also it has many
methods to do user manipulation. Search, delete, add, are some of the
features of the database. 
The program now checks if there is a collision between two email
accounts and handles it by throwing the second user out of the system
(not really, it just blocks registration). 
A minor update in the GUI is that the tooltips look cooler, and latter
buttons, and fonts will.